+++ packages/expo-ui/CHANGELOG.md
@@ -9,6 +9,7 @@
 - [iOS] Added `ShareLink` component. ([#37125](https://github.com/expo/expo/pull/37125)) by [@pchalupa](https://github.com/pchalupa)
 - Add ios `glass` button variant. ([#37373](https://github.com/expo/expo/pull/37373) by [@aleqsio](https://github.com/aleqsio))
 - Added `jetpack-compose-primitives`. ([#36257](https://github.com/expo/expo/pull/36257) by [@kudo](https://github.com/kudo))
+- [android] Added `style` prop for text view and text input view. ([#37628](https://github.com/expo/expo/pull/37628) by [@mateoguzmana](https://github.com/mateoguzmana))
 
 ### 🐛 Bug fixes